+++ packages/ice-plugin-modular-import/index.js
@@ -1,0 +1,21 @@
+module.exports = async ({ chainWebpack }, pluginOptions = []) => {
+  chainWebpack((config) => {
+    ['jsx', 'tsx'].forEach((rule) => {
+      config.module
+        .rule(rule)
+        .use('babel-loader')
+        .tap((options) => {
+          pluginOptions.forEach((item = {}) => {
+            options.plugins.push(
+              [
+                require.resolve('babel-plugin-import'),
+                item,
+                item.libraryName,
+              ]
+            );
+          });
+          return options;
+        });
+    });
+  });
+};