+++ tools/ice-scripts/lib/commands/dev.js
@@ -7,6 +7,7 @@ process.env.NODE_ENV = 'development';
 const chalk = require('chalk');
 const clearConsole = require('react-dev-utils/clearConsole');
 const formatWebpackMessages = require('react-dev-utils/formatWebpackMessages');
+const { collectDetail } = require('@alifd/fusion-collector');
 const webpack = require('webpack');
 const webpackDevMock = require('webpack-dev-mock');
 const WebpackDevServer = require('webpack-dev-server');
@@ -19,13 +20,14 @@ const log = require('../utils/log');
 const checkDepsInstalled = require('../utils/checkDepsInstalled');
 
 module.exports = async function (context, subprocess) {
-  const { applyHook, commandArgs, rootDir, webpackConfig } = context;
+  const { applyHook, commandArgs, rootDir, webpackConfig, pkg } = context;
 
   goldlog('version', {
     version: pkgData.version,
   });
   goldlog('dev', commandArgs);
   log.verbose('dev cliOptions', commandArgs);
+
   await applyHook('beforeDev');
 
   // 与 iceworks 客户端通信
@@ -38,9 +40,21 @@ module.exports = async function (context, subprocess) {
 
   const installedDeps = checkDepsInstalled(rootDir);
   if (!installedDeps) {
-    log.error('项目依赖未安装，请先安装依赖。');
-    process.exit(1);
-    return;
+    return Promise.reject(new Error('项目依赖未安装，请先安装依赖。'));
+  }
+
+  if (!pkg.componentConfig && !pkg.blockConfig) {
+    // only collect project
+    try {
+      collectDetail({
+        rootDir,
+        kit: 'ice-scripts',
+        kitVersion: pkgData.version,
+        cmd_type: 'dev',
+      });
+    } catch (err) {
+      log.warn('collectDetail error', err);
+    }
   }
 
   const HOST = commandArgs.host || '0.0.0.0';
@@ -86,6 +100,8 @@ module.exports = async function (context, subprocess) {
       openBrowser(urls.localUrlForBrowser);
     }
 
+    applyHook('afterDev', stats);
+
     console.log(
       stats.toString({
         colors: true,
@@ -108,38 +124,7 @@ module.exports = async function (context, subprocess) {
           `Compiled successfully in ${(json.time / 1000).toFixed(1)}s!`
         );
       }
-    }
-    applyHook('afterDev', stats);
 
-    if (messages.errors.length) {
-      if (messages.errors.length > 1) {
-        messages.errors.length = 1;
-      }
-      log.error('Failed to compile.\n');
-      console.log(messages.errors.join('\n\n'));
-    } else if (messages.warnings.length) {
-      log.warn('Compiled with warnings.');
-      console.log();
-      messages.warnings.forEach((message) => {
-        console.log(message);
-        console.log();
-      });
-      // Teach some ESLint tricks.
-      console.log('You may use special comments to disable some warnings.');
-      console.log(
-        `Use ${chalk.yellow(
-          '// eslint-disable-next-line'
-        )} to ignore the next line.`
-      );
-      console.log(
-        `Use ${chalk.yellow(
-          '/* eslint-disable */'
-        )} to ignore all warnings in a file.`
-      );
-      console.log();
-    }
-
-    if (isSuccessful) {
       // 服务启动完成切没有任务错误与警告
       send({
         action: 'update_project',
@@ -150,6 +135,33 @@ module.exports = async function (context, subprocess) {
         },
       });
     } else {
+      if (messages.errors.length) {
+        if (messages.errors.length > 1) {
+          messages.errors.length = 1;
+        }
+        console.log(messages.errors.join('\n\n'));
+      } else if (messages.warnings.length) {
+        log.warn('Compiled with warnings.');
+        console.log();
+        messages.warnings.forEach((message) => {
+          console.log(message);
+          console.log();
+        });
+        // Teach some ESLint tricks.
+        console.log('You may use special comments to disable some warnings.');
+        console.log(
+          `Use ${chalk.yellow(
+            '// eslint-disable-next-line'
+          )} to ignore the next line.`
+        );
+        console.log(
+          `Use ${chalk.yellow(
+            '/* eslint-disable */'
+          )} to ignore all warnings in a file.`
+        );
+        console.log();
+      }
+
       // 服务启动完成切没有任务错误与警告
       send({
         action: 'update_project',
@@ -182,7 +194,6 @@ module.exports = async function (context, subprocess) {
   });
 
   devServer.listen(PORT, HOST, (err) => {
-    // 端口被占用，退出程序
     if (err) {
       send({
         action: 'update_project',
@@ -194,6 +205,7 @@ module.exports = async function (context, subprocess) {
       console.error(err);
       process.exit(500);
     } else {
+      applyHook('afterDevServer', devServer);
       send({
         action: 'update_project',
         message: 'server_success',